ETL_STREAMING/requirements.txt
websocket-client==1.7.0
pandas==2.2.1
psycopg2-binary==2.9.9